requirements.txt
streamlit==1.29.0
PyPDF2==3.0.1
python-docx==0.8.11
pandas==2.1.4
numpy==1.24.3
plotly==5.17.0
wordcloud==1.9.2
textstat==0.7.3
spacy==3.7.2
scikit-learn==1.3.2
nltk==3.8.1
requests==2.31.0
python-dotenv==1.0.0
streamlit-chat==0.1.1
streamlit-option-menu==0.3.6
